requirements.txt
discord.py>=2.0
openai
